Rename and move IamRoleResolver
